Move test library code to separate module
