+++ src/tests/lib/curses.py
