+++ src/tests/lib/screen_test_runner.py
@@ -6,9 +6,9 @@ import curses
 import sys
 import os
 
-from tests.key_bindings_for_test import KeyBindingsForTest
-from tests.curses_for_test import CursesForTest
-from tests.screen_for_test import ScreenForTest
+from tests.lib.key_bindings import KeyBindingsForTest
+from tests.lib.curses import CursesForTest
+from tests.lib.screen import ScreenForTest
 from pathpicker.screen_flags import ScreenFlags
 import process_input
 import choose
