+++ src/tests/test_parsing.py
@@ -7,7 +7,7 @@ import unittest
 import os
 
 from pathpicker import parse
-from tests.local_test_cases import LOCAL_TEST_CASES
+from tests.lib.local_test_cases import LOCAL_TEST_CASES
 from pathpicker.formatted_text import FormattedText
 from pathpicker import format
